#168 fix NPE on certain tile entities
