+++ build.gradle
@@ -14,7 +14,7 @@ buildscript {
 }
 apply plugin: 'net.minecraftforge.gradle.forge'
 
-version = "${mc_version}-1.8.7"
+version = "${mc_version}-1.8.8"
 group = "com.lothrazar.cyclicmagic" 
 archivesBaseName = "Cyclic"
