+++ src/main/resources/mcmod.info
@@ -4,7 +4,7 @@
     "modid": "cyclicmagic",
     "name": "Cyclic",
     "description": "Random things that are useful and fun, including powerful items, inventory features, world generation, a few machines, gameplay tweaks, and more. See config to disable absolutely anything.  Report issues to https://github.com/PrinceOfAmber/Cyclic/issues ",
-    "version": "1.8.7",
+    "version": "1.8.8",
     "mcversion": "1.10.2",
     "url": "http://minecraft.curseforge.com/projects/cyclic",
     "updateUrl": "http://minecraft.curseforge.com/projects/cyclic",
